+++ lib/__tests__/mocks/expo-constants.ts
@@ -1,0 +1,4 @@
+export default {
+  appOwnership: 'expo',
+  expoConfig: {},
+};